feat: implement AI command execution in a new module, separating concerns from configuration handling
