+++ src/cli/pr-review.ts
@@ -1,7 +1,8 @@
 import { $ } from 'zx'
 import { Command } from 'commander'
 import { checkGitHubCLI } from '../utils.js'
-import { executeAICommand, loadConfig } from '../config.js'
+import { loadConfig } from '../config.js'
+import { executeAICommand } from '../ai-executor.js'
 
 interface PRInfo {
   url: string
